Allow character builder to hot reload properly
The loading state was defaulted to true, so when it was hot reloaded it
would reset the state to the default. Normally, the componentDidMount
method would cause the loading state to be updated to false, allowing
normal rendering, but hot reload didn't have to remount the component,
so shit got fucked up.